fixed group post comments, display group name with poster name
