+++ src/features/post/PostComponent.tsx
@@ -103,6 +103,7 @@ const PostComponent =  ({ shouldUpdateLikes, post, leaveComment }:
         <Card id="postCard">
             <Card.Header>
                 {/* Link to the poster's profile in Reverb*/}
+                {post.groupName && <Card.Subtitle id="group"><Link to={`group/${post.groupName}`}>{`${post.groupName} / `}</Link></Card.Subtitle>}
                 <Card.Subtitle id="postAuthor"><Link to={`profile/${authorProfile.id}`}>{"" + authorProfile.first_name} {"" + authorProfile.last_name}</Link></Card.Subtitle>
                 {/*Date that the post was made.*/}
                 <Card.Text id = "postTime">{"" + new Date(post.date + 'Z').toLocaleString() }</Card.Text>
